+++ tests/application_modules/test_train_model.py
@@ -40,7 +40,13 @@ def test_integration_test(muteable_test_config: FullConfigSchema):
 def test_crossvalidation(muteable_test_config: FullConfigSchema):
     """Test crossvalidation."""
     cfg = muteable_test_config
-    cfg.train.n_splits = 2
+    train_model(cfg)
+
+
+def test_train_val_predict(muteable_test_config: FullConfigSchema):
+    """Test train without crossvalidation."""
+    cfg = muteable_test_config
+    cfg.data.splits_for_evaluation = ["test"]
     train_model(cfg)